refactor(frontend): move the source location next to the dbt target fields
The location inputs rendered after the advanced-options toggle, which put
them outside the section they belong to. They now sit in the same always
visible block that shows the dataset or schema on the project form, so a
source's location reads as the per-source counterpart of it.

Kept out of the advanced section on purpose: an unset location is not an
error, it silently reads the project's own location, so hiding the field
hides the cause.
