+++ packages/frontend/src/components/ProjectConnection/ProjectFormProvider.tsx
@@ -3,9 +3,9 @@ import Context, { type ProjectFormContext } from './context';
 
 export const ProjectFormProvider: FC<
     React.PropsWithChildren<ProjectFormContext>
-> = ({ savedProject, isDbtSource, children }) => {
+> = ({ savedProject, isDbtSource, projectUuid, children }) => {
     return (
-        <Context.Provider value={{ savedProject, isDbtSource }}>
+        <Context.Provider value={{ savedProject, isDbtSource, projectUuid }}>
             {children}
         </Context.Provider>
     );
